refactor: restructure CLI components and hooks for improved organization and functionality
